sched/walt: Fix pipeline fastpath during active load balance
Pipeline fastpath is skipped when a cpu is already running a pipeline
task. It may happen that the current pipeline task and the task for
which we are finding the best cpu are same. This scenario arises during
load balancing on every scheduler tick, where we find the new best cpu
for the current task. Fix this by allowing pipeline fastpath if current
pipeline process on a cpu is same as the process for which we are
finding the best cpu.

Change-Id: Iaf935694c7c580b2e191793491130c1a3ccbedc4
Signed-off-by: Atul Pant <quic_atulpant@quicinc.com>
(cherry picked from commit 0f8077d6550326ba474afcc0282fa86c9cf26753)